requirements.txt
aiogram == 3.18.0
python_dotenv == 1.0.1